+++ packages/create-capacitor-svelte-app/src/creator.js
@@ -66,7 +66,7 @@ export async function createIonicSvelte(opts) {
 	}
 
 	if (!(opts?.quiet)) {
-		console.log('Working: Creating base Svelte Kit install supercharged with Ionic.');
+		console.log('Working: Creating base Svelte Kit install supercharged with Capacitor.');
 	}
 	fs.mkdirp(opts.path);
 
@@ -217,6 +217,28 @@ export async function createIonicSvelte(opts) {
 				console.warn('TSconfig read/write error - ', e);
 			}
 		}
+
+		// hot reload support - change the vite build script and point server url
+		try {
+			const packagagejson = fs.readFileSync('package.json', 'utf-8');
+			//	console.log('Reading tsconfig ', tsconfig);
+			const packagagejsonnew = packagagejson.replace('"dev": "vite dev"', `"dev": "vite dev --host"`);
+
+			//	console.log('New tsconfig ', tsconfignew);
+			out(path.resolve(process.cwd(), './', 'package.json'), packagagejsonnew)
+		} catch (e) {
+			console.warn('TSconfig read/write error - ', e);
+		}
+
+		out(
+			'capacitor.config.json',
+			`{
+				"server": {
+				  "url": "http://192.168.137.1:5173/",
+				  "cleartext": true
+				}
+			}`
+		);
 	}
 
 	return opts;